openssl.c: cast to expected OpenSSL API types

+++ src/openssl.c
@@ -119,29 +119,29 @@ _libssh2_rsa_new(libssh2_rsa_ctx ** rsa,
     BIGNUM * iqmp = 0;
 
     e = BN_new();
-    BN_bin2bn(edata, elen, e);
+    BN_bin2bn(edata, (int) elen, e);
 
     n = BN_new();
-    BN_bin2bn(ndata, nlen, n);
+    BN_bin2bn(ndata, (int) nlen, n);
 
     if(ddata) {
         d = BN_new();
-        BN_bin2bn(ddata, dlen, d);
+        BN_bin2bn(ddata, (int) dlen, d);
 
         p = BN_new();
-        BN_bin2bn(pdata, plen, p);
+        BN_bin2bn(pdata, (int) plen, p);
 
         q = BN_new();
-        BN_bin2bn(qdata, qlen, q);
+        BN_bin2bn(qdata, (int) qlen, q);
 
         dmp1 = BN_new();
-        BN_bin2bn(e1data, e1len, dmp1);
+        BN_bin2bn(e1data, (int) e1len, dmp1);
 
         dmq1 = BN_new();
-        BN_bin2bn(e2data, e2len, dmq1);
+        BN_bin2bn(e2data, (int) e2len, dmq1);
 
         iqmp = BN_new();
-        BN_bin2bn(coeffdata, coefflen, iqmp);
+        BN_bin2bn(coeffdata, (int) coefflen, iqmp);
     }
 
     *rsa = RSA_new();
@@ -210,8 +210,9 @@ _libssh2_rsa_sha2_verify(libssh2_rsa_ctx * rsactx,
         return -1; /* failure */
     }
 
-    ret = RSA_verify(nid_type, hash, hash_len,
-                     (unsigned char *) sig, sig_len, rsactx);
+    ret = RSA_verify(nid_type, hash, (unsigned int) hash_len,
+                     (unsigned char *) sig,
+                     (unsigned int) sig_len, rsactx);
 
     free(hash);
 
@@ -248,20 +249,20 @@ _libssh2_dsa_new(libssh2_dsa_ctx ** dsactx,
     BIGNUM * priv_key = NULL;
 
     p_bn = BN_new();
-    BN_bin2bn(p, p_len, p_bn);
+    BN_bin2bn(p, (int) p_len, p_bn);
 
     q_bn = BN_new();
-    BN_bin2bn(q, q_len, q_bn);
+    BN_bin2bn(q, (int) q_len, q_bn);
 
     g_bn = BN_new();
-    BN_bin2bn(g, g_len, g_bn);
+    BN_bin2bn(g, (int) g_len, g_bn);
 
     pub_key = BN_new();
-    BN_bin2bn(y, y_len, pub_key);
+    BN_bin2bn(y, (int) y_len, pub_key);
 
     if(x_len) {
         priv_key = BN_new();
-        BN_bin2bn(x, x_len, priv_key);
+        BN_bin2bn(x, (int) x_len, priv_key);
     }
 
     *dsactx = DSA_new();
@@ -426,17 +427,17 @@ _libssh2_ecdsa_verify(libssh2_ecdsa_ctx * ctx,
     BIGNUM *pr = BN_new();
     BIGNUM *ps = BN_new();
 
-    BN_bin2bn(r, r_len, pr);
-    BN_bin2bn(s, s_len, ps);
+    BN_bin2bn(r, (int) r_len, pr);
+    BN_bin2bn(s, (int) s_len, ps);
     ECDSA_SIG_set0(ecdsa_sig, pr, ps);
 
 #else
     ECDSA_SIG ecdsa_sig_;
     ECDSA_SIG *ecdsa_sig = &ecdsa_sig_;
     ecdsa_sig_.r = BN_new();
-    BN_bin2bn(r, r_len, ecdsa_sig_.r);
+    BN_bin2bn(r, (int) r_len, ecdsa_sig_.r);
     ecdsa_sig_.s = BN_new();
-    BN_bin2bn(s, s_len, ecdsa_sig_.s);
+    BN_bin2bn(s, (int) s_len, ecdsa_sig_.s);
 #endif
 
     if(type == LIBSSH2_EC_CURVE_NISTP256) {
@@ -488,9 +489,9 @@ _libssh2_cipher_crypt(_libssh2_cipher_ctx * ctx,
     (void) encrypt;
 
 #ifdef HAVE_OPAQUE_STRUCTS
-    ret = EVP_Cipher(*ctx, buf, block, blocksize);
+    ret = EVP_Cipher(*ctx, buf, block, (unsigned int) blocksize);
 #else
-    ret = EVP_Cipher(ctx, buf, block, blocksize);
+    ret = EVP_Cipher(ctx, buf, block, (unsigned int) blocksize);
 #endif
 
 #if (defined(OPENSSL_VERSION_MAJOR) && OPENSSL_VERSION_MAJOR >= 3) || \
@@ -536,7 +537,7 @@ void _libssh2_openssl_crypto_exit(void)
 static int
 passphrase_cb(char *buf, int size, int rwflag, char *passphrase)
 {
-    int passphrase_len = strlen(passphrase);
+    int passphrase_len = (int) strlen(passphrase);
     (void) rwflag;
 
     if(passphrase_len > (size - 1)) {
@@ -562,7 +563,7 @@ read_private_key_from_memory(void **key_ctx,
 
     *key_ctx = NULL;
 
-    bp = BIO_new_mem_buf((char *)filedata, filedata_len);
+    bp = BIO_new_mem_buf((char *)filedata, (int)filedata_len);
     if(!bp) {
         return -1;
     }
@@ -2013,11 +2014,14 @@ _libssh2_rsa_sha2_sign(LIBSSH2_SESSION * session,
     }
 
     if(hash_len == SHA_DIGEST_LENGTH)
-        ret = RSA_sign(NID_sha1, hash, hash_len, sig, &sig_len, rsactx);
+        ret = RSA_sign(NID_sha1,
+                       hash, (unsigned int) hash_len, sig, &sig_len, rsactx);
     else if(hash_len == SHA256_DIGEST_LENGTH)
-        ret = RSA_sign(NID_sha256, hash, hash_len, sig, &sig_len, rsactx);
+        ret = RSA_sign(NID_sha256,
+                       hash, (unsigned int) hash_len, sig, &sig_len, rsactx);
     else if(hash_len == SHA512_DIGEST_LENGTH)
-        ret = RSA_sign(NID_sha512, hash, hash_len, sig, &sig_len, rsactx);
+        ret = RSA_sign(NID_sha512,
+                       hash, (unsigned int) hash_len, sig, &sig_len, rsactx);
     else {
         _libssh2_error(session, LIBSSH2_ERROR_PROTO,
                        "Unsupported hash digest length");
@@ -2108,7 +2112,7 @@ _libssh2_ecdsa_sign(LIBSSH2_SESSION * session, libssh2_ecdsa_ctx * ec_ctx,
     unsigned char *temp_buffer = NULL;
     unsigned char *out_buffer = NULL;
 
-    ECDSA_SIG *sig = ECDSA_do_sign(hash, hash_len, ec_ctx);
+    ECDSA_SIG *sig = ECDSA_do_sign(hash, (int) hash_len, ec_ctx);
     if(sig == NULL)
         return -1;
 #ifdef HAVE_OPAQUE_STRUCTS
@@ -2596,7 +2600,7 @@ gen_publickey_from_ecdsa_openssh_priv_data(LIBSSH2_SESSION *session,
         goto fail;
     }
 
-    BN_bin2bn(exponent, exponentlen, bn_exponent);
+    BN_bin2bn(exponent, (int) exponentlen, bn_exponent);
     rc = (EC_KEY_set_private_key(ec_key, bn_exponent) != 1);
 
     if(rc == 0 && ec_key != NULL && pubkeydata != NULL && method != NULL) {
@@ -3070,7 +3074,7 @@ _libssh2_ecdh_gen_k(_libssh2_bn **k, _libssh2_ec_key *private_key,
         goto clean_exit;
     }
 
-    BN_bin2bn(secret, secret_len, *k);
+    BN_bin2bn(secret, (int) secret_len, *k);
 
 clean_exit:
 
@@ -3701,7 +3705,7 @@ _libssh2_pub_priv_keyfilememory(LIBSSH2_SESSION *session,
                    LIBSSH2_TRACE_AUTH,
                    "Computing public key from private key."));
 
-    bp = BIO_new_mem_buf((char *)privatekeydata, privatekeydata_len);
+    bp = BIO_new_mem_buf((char *)privatekeydata, (int)privatekeydata_len);
     if(!bp)
         return _libssh2_error(session, LIBSSH2_ERROR_ALLOC,
                               "Unable to allocate memory when"
@@ -3791,7 +3795,7 @@ _libssh2_sk_pub_keyfilememory(LIBSSH2_SESSION *session,
                    LIBSSH2_TRACE_AUTH,
                    "Computing public key from private key."));
 
-    bp = BIO_new_mem_buf((char *)privatekeydata, privatekeydata_len);
+    bp = BIO_new_mem_buf((char *)privatekeydata, (int)privatekeydata_len);
     if(!bp)
         return _libssh2_error(session, LIBSSH2_ERROR_ALLOC,
                               "Unable to allocate memory when"
